Move standalone widgets into separate subpackge
