+++ naturtag/app/image_controller.py
@@ -6,9 +6,9 @@ from PySide6.QtGui import QIntValidator
 from PySide6.QtWidgets import QApplication, QGroupBox, QLabel, QLineEdit, QWidget
 
 from naturtag.app.image_gallery import ImageGallery
-from naturtag.app.layouts import HorizontalLayout, VerticalLayout
 from naturtag.metadata.inat_metadata import get_ids_from_url, tag_images
 from naturtag.settings import Settings
+from naturtag.widgets import HorizontalLayout, VerticalLayout
 
 logger = getLogger(__name__)
