+++ naturtag/metadata/__init__.py
@@ -5,4 +5,4 @@
 from naturtag.metadata.image_metadata import ImageMetadata
 from naturtag.metadata.keyword_metadata import HIER_KEYWORD_TAGS, KEYWORD_TAGS, KeywordMetadata
 from naturtag.metadata.meta_metadata import MetaMetadata
-from naturtag.metadata.inat_metadata import get_inat_metadata, tag_image, tag_images
+from naturtag.metadata.inat_metadata import INAT_CLIENT, get_inat_metadata, tag_image, tag_images
