+++ naturtag/widgets/__init__.py
@@ -1,0 +1,14 @@
+# flake8: noqa: F401
+# isort: skip_file
+from naturtag.widgets.layouts import (
+    ClearMixin,
+    FlowLayout,
+    HorizontalLayout,
+    StylableWidget,
+    StyleMixin,
+    VerticalLayout,
+)
+from naturtag.widgets.autocomplete import TaxonAutocomplete
+from naturtag.widgets.images import IconLabel, PixmapLabel
+from naturtag.widgets.logger import QtRichHandler, init_handler
+from naturtag.widgets.toggle_switch import ToggleSwitch
